+++ changelog.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- ```getOrCreate``` method to the Container.
+
 ### Fixed
 - Overwriting and removing services from the container.
+- Merged Container and Injector.
+- Removed namespace prefix feature.
 
 ## [0.1.0] - 2018-07-29
 ### Added